# feedback-back/src/main/resources/api.yaml
openapi: 3.0.0
...
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/KevinTAN52/FeedBackISEP/1.0.0
info:
  description: |
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: "1.0.0"
  title: FeedBack
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/conversations':
    post:
      tags:
        - messages
      summary: Creates a conversation
      description: Creates a conversation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation:
                  $ref: '#/components/schemas/Conversation'
                message:
                  $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    $ref: '#/components/schemas/Conversation'
                  message:
                    $ref: '#/components/schemas/Message'
            application/xml:
              schema:
                type: object
                properties:
                  conversation:
                    $ref: '#/components/schemas/Conversation'
                  message:
                    $ref: '#/components/schemas/Message'
        '404':
          description: User not found
    get:
      tags:
        - messages
      summary: Get user's conversations
      description: Get user's conversations
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '404':
          description: User not found
  '/conversations/{conversationId}':
    delete:
      tags:
        - messages
      summary: Deletes a conversation
      description: Deletes a conversation
      parameters:
        - name: conversationId
          in: path
          description: id of the conversation to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
        '404':
          description:  Conversation not found
  '/conversations/{conversationId}/messages':
    post:
      tags:
        - messages
      summary: Reply to a conversation
      description: Reply to a conversation
      parameters:
        - name: conversationId
          in: path
          description: id of the conversation to reply to
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
            application/xml:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description:  Message not found
    get:
      tags:
        - messages
      summary: Get user's messages of a conversation
      description: Get user's messages of a conversation
      parameters:
        - name: conversationId
          in: path
          description: id of the conversation to reply to
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '404':
          description:  conversation not found
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
        conversation_id:
          type: integer
          format: int64
        is_from_sender:
          type: boolean
        content:
          type: string
        read:
          type: boolean
        dateTime:
          type: string
    Conversation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        from_id:
          type: integer
          format: int64
        to_id:
          type: integer
          format: int64
        subject:
          type: string
        status:
          type: string
        dateTime:
          type: string
        messages_id:
          type: string
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header